Fixed bot from looping itself
remove the ability for the bot to loop on itself if a user banned a word that the bot used. i.e. banning the word Obama, the bot would loop on itself if a user were to then type Obama, because the bot uses the word Obama in it's warning message for using a banned word.